yey models

+++ src/main/java/net/quantumfusion/dash/util/TimeHelper.java
@@ -6,6 +6,6 @@ import java.time.Instant;
 public class TimeHelper {
 
     public static float getDecimalMs(Instant start, Instant stop) {
-        return Math.round(Duration.between(start, stop).getNano() / 100000f) / 10f;
+        return Math.round(Duration.between(start, stop).getNano() / (double)100000) / 10f;
     }
 }
